refactor: move more template code into mixin
