+++ openassessment/xblock/openassessmentblock.py
@@ -51,8 +51,7 @@ from openassessment.xblock.workflow_mixin import WorkflowMixin
 from openassessment.xblock.team_workflow_mixin import TeamWorkflowMixin
 from openassessment.xblock.openassesment_template_mixin import OpenAssessmentTemplatesMixin
 from openassessment.xblock.xml import parse_from_xml, serialize_content_to_xml
-from openassessment.xblock.editor_config import AVAILABLE_EDITORS
-from openassessment.xblock.load_static import LoadStatic
+
 
 from openassessment.xblock.ui_mixins.legacy.mixin import LegacyViewUIMixin
 
@@ -64,45 +63,6 @@ from openassessment.xblock.assessments.staff_assessment import StaffAssessmentAP
 from openassessment.xblock.assessments.student_training import StudentTrainingAPI
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
-
-
-UI_MODELS = {
-    "submission": {
-        "name": "submission",
-        "class_id": "step--response",
-        "title": "Your Response"
-    },
-    "student-training": {
-        "name": "student-training",
-        "class_id": "step--student-training",
-        "title": "Learn to Assess Responses"
-    },
-    "peer-assessment": {
-        "name": "peer-assessment",
-        "class_id": "step--peer-assessment",
-        "title": "Assess Peers"
-    },
-    "self-assessment": {
-        "name": "self-assessment",
-        "class_id": "step--self-assessment",
-        "title": "Assess Your Response"
-    },
-    "staff-assessment": {
-        "name": "staff-assessment",
-        "class_id": "step--staff-assessment",
-        "title": "Staff Grade"
-    },
-    "grade": {
-        "name": "grade",
-        "class_id": "step--grade",
-        "title": "Your Grade:"
-    },
-    "leaderboard": {
-        "name": "leaderboard",
-        "class_id": "step--leaderboard",
-        "title": "Top Responses"
-    }
-}
 
 
 def load(path):
@@ -712,58 +672,6 @@ class OpenAssessmentBlock(
             additional_js_context=context_dict,
         )
 
-    def _create_fragment(
-        self,
-        template,
-        context_dict,
-        initialize_js_func,
-        additional_css=None,
-        additional_js=None,
-        additional_js_context=None
-    ):
-        """
-        Creates a fragment for display.
-
-        """
-        fragment = Fragment(template.render(context_dict))
-
-        if additional_css is None:
-            additional_css = []
-        if additional_js is None:
-            additional_js = []
-
-        i18n_service = self.runtime.service(self, 'i18n')
-        if hasattr(i18n_service, 'get_language_bidi') and i18n_service.get_language_bidi():
-            css_url = LoadStatic.get_url("openassessment-rtl.css")
-        else:
-            css_url = LoadStatic.get_url("openassessment-ltr.css")
-
-        # TODO: load CSS and JavaScript as URLs once they can be served by the CDN
-        for css in additional_css:
-            fragment.add_css_url(css)
-        fragment.add_css_url(css_url)
-
-        # minified additional_js should be already included in 'make javascript'
-        fragment.add_javascript_url(LoadStatic.get_url("openassessment-lms.js"))
-
-        js_context_dict = {
-            "ALLOWED_IMAGE_MIME_TYPES": self.ALLOWED_IMAGE_MIME_TYPES,
-            "ALLOWED_FILE_MIME_TYPES": self.ALLOWED_FILE_MIME_TYPES,
-            "FILE_EXT_BLACK_LIST": self.FILE_EXT_BLACK_LIST,
-            "FILE_TYPE_WHITE_LIST": self.white_listed_file_types,
-            "MAXIMUM_FILE_UPLOAD_COUNT": self.MAX_FILES_COUNT,
-            "TEAM_ASSIGNMENT": self.is_team_assignment(),
-            "AVAILABLE_EDITORS": AVAILABLE_EDITORS,
-            "TEXT_RESPONSE_EDITOR": self.text_response_editor,
-        }
-        # If there's any additional data to be passed down to JS
-        # include it in the context dict
-        if additional_js_context:
-            js_context_dict.update({"CONTEXT": additional_js_context})
-
-        fragment.initialize_js(initialize_js_func, js_context_dict)
-        return fragment
-
     @property
     def is_admin(self):
         """
@@ -825,36 +733,6 @@ class OpenAssessmentBlock(
         if hasattr(self, 'xmodule_runtime'):
             return self.xmodule_runtime.get_real_user is not None  # pylint: disable=no-member
         return False
-
-    def _create_ui_models(self):
-        """Combine UI attributes and XBlock configuration into a UI model.
-
-        This method takes all configuration for this XBlock instance and appends
-        UI attributes to create a UI Model for rendering all assessment modules.
-        This allows a clean separation of static UI attributes from persistent
-        XBlock configuration.
-
-        """
-        ui_models = [UI_MODELS["submission"]]
-        staff_assessment_required = False
-        for assessment in self.valid_assessments:
-            if assessment["name"] == "staff-assessment":
-                if not assessment["required"]:
-                    continue
-                staff_assessment_required = True
-            ui_model = UI_MODELS.get(assessment["name"])
-            if ui_model:
-                ui_models.append(dict(assessment, **ui_model))
-
-        if not staff_assessment_required and self.staff_assessment_exists(self.submission_uuid):
-            ui_models.append(UI_MODELS["staff-assessment"])
-
-        ui_models.append(UI_MODELS["grade"])
-
-        if self.leaderboard_show > 0 and not self.teams_enabled:
-            ui_models.append(UI_MODELS["leaderboard"])
-
-        return ui_models
 
     @staticmethod
     def workbench_scenarios():
